nav. Refactor NavBar, fix in view detection and blocked scroll on phones. Fix smaller issues
